Fixing bugs
Added an exception catcher to the download statement so downloads with errors didn't crash it, just moved on.

Added any erroneous downloads to a list to be printed on completion for manual perusal

Seperated the loading of data into its own function rather than duplicating it.

Added the ability to load API key data from a text file rather than having it hardcoded.  Saves me remembering to strip it out each time I try to distribute the script.  It can still be hardcoded into the script by the user, but if it isn't, it will look for the apikey.txt file.